backend/requirements.txt
neo4j
bs4
pdfplumber
langchain
requests
faiss-cpu
numpy